+++ src/eyconf/validation/_to_json.py
@@ -14,11 +14,11 @@ from typing import (
 )
 
 from jsonschema import Draft202012Validator
-from typing_extensions import NotRequired
+from typing_extensions import Final, NotRequired
 
 __all__ = ["to_json_schema", "primitives"]
 
-primitives: dict[type[object], "str"] = {
+primitives: Final[dict[type[object], "str"]] = {
     str: "string",
     int: "integer",
     float: "number",
@@ -28,7 +28,11 @@ primitives: dict[type[object], "str"] = {
 
 
 @lru_cache(maxsize=None)
-def to_json_schema(type: type, check_schema: bool = True) -> dict:
+def to_json_schema(
+    type: type,
+    check_schema: bool = True,
+    allow_additional: bool = True,
+) -> dict:
     """Convert a TypedDict or dataclass to a JSON schema.
 
     Parameters
@@ -37,6 +41,8 @@ def to_json_schema(type: type, check_schema: bool = True) -> dict:
         The TypedDict or dataclass to convert.
     check_schema : bool
         Whether to check the schema for validity.
+    allow_additional : bool
+        Whether to allow extra, unrecognized properties in the schema.
 
     Raises
     ------
@@ -49,6 +55,7 @@ def to_json_schema(type: type, check_schema: bool = True) -> dict:
         "type": "object",
         "properties": {},
         "required": [],
+        "additionalProperties": allow_additional,
     }
 
     # Get type hints for the TypedDict
@@ -56,7 +63,7 @@ def to_json_schema(type: type, check_schema: bool = True) -> dict:
 
     # Add the type hints to the schema
     for field_name, field_type in type_hints.items():
-        p, r = __convert_type_to_schema(field_type)
+        p, r = __convert_type_to_schema(field_type, allow_additional=allow_additional)
         schema["properties"][field_name] = p
 
         if r:
@@ -77,6 +84,7 @@ SchemaType = Union[
 
 def __convert_type_to_schema(
     field_type: type,
+    **kwargs,
 ) -> tuple[SchemaType, bool]:
     """
     Convert a type to a JSON schema.
@@ -85,6 +93,8 @@ def __convert_type_to_schema(
     ----------
     field_type : type
         The type to convert.
+    **kwargs : dict
+        Passed recusrively and to `to_json_schema()` if called in here.
 
     Returns
     -------
@@ -109,7 +119,7 @@ def __convert_type_to_schema(
     if origin is NotRequired:
         field_type = get_args(field_type)[0]
         is_required = False
-        return __convert_type_to_schema(field_type)[0], is_required
+        return __convert_type_to_schema(field_type, **kwargs)[0], is_required
 
     # Handler union
     if origin is Union or origin is UnionType:
@@ -127,12 +137,12 @@ def __convert_type_to_schema(
         if len(allowed_types) > 1:
             return {
                 "anyOf": [
-                    __convert_type_to_schema(allowed_type)[0]
+                    __convert_type_to_schema(allowed_type, **kwargs)[0]
                     for allowed_type in allowed_types
                 ]
             }, is_required
         elif len(allowed_types) == 1:
-            t, _ = __convert_type_to_schema(allowed_types.pop())
+            t, _ = __convert_type_to_schema(allowed_types.pop(), **kwargs)
             return t, is_required
         else:
             raise ValueError("Union type must have at least one type!")
@@ -141,13 +151,13 @@ def __convert_type_to_schema(
     if origin in [list, set, tuple, Sequence, ABCSequence]:
         return {
             "type": "array",
-            "items": __convert_type_to_schema(get_args(field_type)[0])[0],
+            "items": __convert_type_to_schema(get_args(field_type)[0], **kwargs)[0],
         }, is_required
 
     # Handle TypedDict and dataclasses
     try:
         if issubclass(field_type, Dict) or is_dataclass(field_type):
-            return to_json_schema(field_type), is_required
+            return to_json_schema(field_type, **kwargs), is_required
     except TypeError:
         # Throws an error in case of a type that is not a class
         pass
@@ -161,7 +171,7 @@ def __convert_type_to_schema(
         return {
             "type": "object",
             "patternProperties": {
-                ".*": __convert_type_to_schema(value_type)[0]
+                ".*": __convert_type_to_schema(value_type, **kwargs)[0]
             },
         }, is_required
 
@@ -178,14 +188,13 @@ def __convert_type_to_schema(
 
 
 def __infer_type_from_values(values: tuple | list):
-
-    types : list[type] = []
+    types: list[type] = []
     for value in values:
         value_type = type(value)
         if value_type not in types:
             types.append(value_type)
 
-    type_names : list[str] = []
+    type_names: list[str] = []
     for t in types:
         if t in primitives:
             type_names.append(primitives[t])