+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 * Now supports patternProperties for nested types, where they key can be arbitrary but its content should still adhere to a specific type.
 * Added support for validating mixed type literals.
+* Added functionality to disallow or allow additional properties in configuration schemas.
+
+## Changed
+
 * The `NoneType` type has been added to to literals and represents `null` values in YAML.
 
 ## [0.1.0] - 2025-08-07